Form: rewrite from 'reduxForm'
