+++ src/utils/helpers.ts
@@ -5,3 +5,7 @@ export function isString(cand: any): cand is string {
 export function isPromise<T>(cand: any): cand is PromiseLike<T> {
   return Boolean(cand) && typeof cand === 'object' && typeof cand.then === 'function';
 }
+
+export function isFunction(cand: any): cand is Function {
+  return typeof cand === 'function';
+}
